Refactor expense hooks to work with cache
