+++ services/frontend/src/state/use-expense/lib/cache.js
@@ -1,0 +1,19 @@
+/**
+ *
+ * @param {query, variables} gql Used to identify the query that needs to be updated
+ */
+
+// It is not a given that the cache was already populated
+export const updateCacheAfterCreate = (gql) => (cache, res) => {
+  try {
+    cache.writeQuery({
+      ...gql,
+      data: {
+        transactions: [
+          ...res.data.insert_expense_transactions.returning,
+          ...cache.readQuery(gql).transactions,
+        ],
+      },
+    });
+  } catch (err) {};
+};
