+++ services/frontend/src/state/use-expense/lib/constants.js
@@ -1,0 +1,1 @@
+export const DEFAULT_OPTIONS = { limit: 20 };
